tsql-core/src/test/java/com/xy/xsql/tsql/converter/functions/conversion/CastConverterTest.java
package com.xy.xsql.tsql.converter.functions.conversion;

import com.xy.xsql.block.core.printer.MetaBlockPrinter;
import com.xy.xsql.block.model.BlockMeta;
import com.xy.xsql.tsql.converter.functions.configuration.DbTsConverter;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2018/5/31.
 */
public class CastConverterTest {

    @Test
    public void testMetaPrint() throws Exception {
        BlockMeta b = CastConverter.meta;

        StringWriter writer = MetaBlockPrinter.print(b);

        System.out.println(writer);
        Assert.assertEquals(writer.toString(),
                "<CAST> ::=\n" +
                        "CAST ( expression AS data_type [ ( length ) ] )");
    }

}